display conditions for schema max 1, reorg test schemas
